microservices/pdftool/pdftool/app.py
from flask import Flask
from routes.routes import initialize_routes
from flask_restful import Api
import os
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

app.config["DEBUG"] = True
api = Api(app)


initialize_routes(api)

if __name__ == "__main__":
    ENVIRONMENT_DEBUG = os.environ.get("APP_DEBUG", True)
    ENVIRONMENT_PORT = os.environ.get("APP_PORT", 5000)
    app.run(host="0.0.0.0", port=ENVIRONMENT_PORT, debug=ENVIRONMENT_DEBUG)
