microservices/pdftool/pdftool/requirements.txt
uwsgi
Flask
flask_restful
mimerender
cleancat
python-dotenv
wheel
bcrypt
uuid
pycurl
requests
awesome-slugify
datetime
pandas
openpyxl
xlsxwriter
bson
python-rclone
Flask-WTF
PyPDF2
argparse
boto3
numpy
pymupdf
